+++ src/main/java/com/cloudogu/scm/tracemonitor/TraceStore.java
@@ -24,69 +24,95 @@
 package com.cloudogu.scm.tracemonitor;
 
 import com.cloudogu.scm.tracemonitor.config.GlobalConfigStore;
+import com.google.common.annotations.VisibleForTesting;
+import com.google.common.util.concurrent.Striped;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.apache.shiro.SecurityUtils;
-import sonia.scm.store.ConfigurationEntryStoreFactory;
 import sonia.scm.store.DataStore;
+import sonia.scm.store.DataStoreFactory;
 import sonia.scm.trace.SpanContext;
 
 import javax.inject.Inject;
+import javax.inject.Singleton;
 import javax.xml.bind.annotation.XmlAccessType;
 import javax.xml.bind.annotation.XmlAccessorType;
 import javax.xml.bind.annotation.XmlElement;
 import javax.xml.bind.annotation.XmlRootElement;
 import java.util.ArrayList;
 import java.util.Collection;
+import java.util.Collections;
 import java.util.List;
-import java.util.stream.Collectors;
+import java.util.concurrent.locks.Lock;
+import java.util.concurrent.locks.ReadWriteLock;
+import java.util.function.Supplier;
 
+@Singleton
 public class TraceStore {
 
   private static final String STORE_NAME = "trace-monitor";
-  private final ConfigurationEntryStoreFactory storeFactory;
+  private final DataStoreFactory storeFactory;
   private final GlobalConfigStore globalConfigStore;
+  private final Striped<ReadWriteLock> locks = Striped.readWriteLock(10);
 
   @Inject
-  public TraceStore(ConfigurationEntryStoreFactory storeFactory, GlobalConfigStore globalConfigStore) {
+  public TraceStore(DataStoreFactory storeFactory, GlobalConfigStore globalConfigStore) {
     this.storeFactory = storeFactory;
     this.globalConfigStore = globalConfigStore;
   }
 
   public Collection<SpanContext> getAll() {
     SecurityUtils.getSubject().checkPermission("traceMonitor:read");
-    DataStore<StoreEntry> store = createStore();
     List<SpanContext> spans = new ArrayList<>();
-    store.getAll().values().forEach(entry -> spans.addAll(entry.getSpans()));
+    createStore()
+      .getAll()
+      .values()
+      .forEach(entry -> spans.addAll(entry.getSpans()));
     return spans;
   }
 
   public Collection<SpanContext> get(String kind) {
-    return getAll().stream()
-      .filter(span -> kind.equalsIgnoreCase(span.getKind()))
-      .collect(Collectors.toList());
+    SecurityUtils.getSubject().checkPermission("traceMonitor:read");
+    return doSynchronized(kind, false, () -> Collections.unmodifiableCollection(createStore().get(kind).getSpans()));
   }
 
-  void add(SpanContext spanContext) {
-    DataStore<StoreEntry> store = createStore();
-    StoreEntry storeEntry = store.get(spanContext.getKind());
-    int configuredStoreSize = globalConfigStore.get().getStoreSize();
-    if (storeEntry == null) {
-      storeEntry = new StoreEntry(configuredStoreSize);
-    }
-    storeEntry = resizeStoreEntryMaxSize(storeEntry, configuredStoreSize);
-    storeEntry.getSpans().add(spanContext);
-    store.put(spanContext.getKind(), storeEntry);
-
+  synchronized void add(SpanContext spanContext) {
+    doSynchronized(spanContext.getKind(), true, () -> {
+      DataStore<StoreEntry> store = createStore();
+      StoreEntry storeEntry = store.get(spanContext.getKind());
+      int configuredStoreSize = globalConfigStore.get().getStoreSize();
+      if (storeEntry == null) {
+        storeEntry = new StoreEntry(configuredStoreSize);
+      }
+      storeEntry = resizeStoreEntryMaxSize(storeEntry, configuredStoreSize);
+      storeEntry.getSpans().add(spanContext);
+      store.put(spanContext.getKind(), storeEntry);
+      return null;
+    });
   }
 
   private StoreEntry resizeStoreEntryMaxSize(StoreEntry storeEntry, int configuredStoreSize) {
     if (storeEntry.getSpans().maxSize != configuredStoreSize) {
-      StoreEntry resizedStoreEntry = new StoreEntry(configuredStoreSize);
-      storeEntry.getSpans().forEach(s -> resizedStoreEntry.getSpans().add(s));
-      storeEntry = resizedStoreEntry;
+      if (configuredStoreSize > storeEntry.getSpans().maxSize) {
+        storeEntry.getSpans().maxSize = configuredStoreSize;
+      } else {
+        StoreEntry resizedStoreEntry = new StoreEntry(configuredStoreSize);
+        storeEntry.getSpans().forEach(s -> resizedStoreEntry.getSpans().add(s));
+        storeEntry = resizedStoreEntry;
+      }
     }
     return storeEntry;
+  }
+
+  private <T> T doSynchronized(String category, boolean write, Supplier<T> callback) {
+    final ReadWriteLock lockFactory = locks.get(category);
+    Lock lock = write ? lockFactory.writeLock() : lockFactory.readLock();
+    lock.lock();
+    try {
+      return callback.get();
+    } finally {
+      lock.unlock();
+    }
   }
 
   private DataStore<StoreEntry> createStore() {
@@ -97,11 +123,12 @@ public class TraceStore {
   @XmlAccessorType(XmlAccessType.FIELD)
   @Getter
   @NoArgsConstructor
-  static class StoreEntry {
+  public static class StoreEntry {
     @XmlElement(name = "request")
     private EvictingQueue<SpanContext> spans;
 
-    StoreEntry(int storeSize) {
+    @VisibleForTesting
+    public StoreEntry(int storeSize) {
       spans = EvictingQueue.create(storeSize);
     }
   }