+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+### Fixed
+- Synchronize store access to prevent concurrent edit errors ([#2](https://github.com/scm-manager/scm-trace-monitor-plugin/pull/2))
+
 ## [1.0.0] - 2020-11-06
 ### Added
 - Add trace monitor and configuration ([#1](https://github.com/scm-manager/scm-trace-monitor-plugin/pull/1))